-Makefile is now more intelligent about what to build, also assumes that i686-elf-g++ is in your path now.\n-Terminal now handles newlines correctly

+++ create_iso.sh
@@ -1,4 +1,12 @@
-mkdir -p isodir/boot/grub
-cp myos.bin isodir/boot/myos.bin
-cp grub.cfg isodir/boot/grub/grub.cfg
-grub-mkrescue -o myos.iso isodir
+last_modified_os=`stat -c "%Y" myos.bin`
+last_modified_iso=`stat -c "%Y" myos.iso`
+
+if [ ! -f myos.iso ] || [ $last_modified_iso -lt $last_modified_os ]
+then
+	mkdir -p isodir/boot/grub
+	cp myos.bin isodir/boot/myos.bin
+	cp grub.cfg isodir/boot/grub/grub.cfg
+	grub-mkrescue -o myos.iso isodir	
+else
+	echo "iso allready up to date"
+fi
